MVP
Creates basic functionality with some typing and error handling.

This cli app is meant to search a directory and all it's sub-directories for SRT formatted subtitle files and then search each file for any ads within the files. The search method for finding the ads in the subtitle text are very rough at the moment and only looks for the string OpenSubtitles (because conceivably there's no reason for this to be real dialogue in a movie) and for any subtitles that try to do any font or color overloading because ime this is reserved for plugging the subtitle maker's social media accounts.